refactor: satisfy repository shape contracts
