+++ cli-rs/src/agent/adapter/mod.rs
@@ -1,0 +1,29 @@
+use crate::agent;
+use crate::cli::{
+    AgentCallsArgs, AgentCommand, AgentDiagnosticsArgs, AgentDiagnosticsViewArgs,
+    AgentHierarchyArgs, AgentHierarchyDirection, AgentImpactArgs, AgentImpactViewArgs,
+    AgentImplementationsArgs, AgentNativeGraphArgs, AgentReferencesArgs, AgentRelationDepth,
+    AgentRelationLimit, AgentRelationViewArgs, AgentReusableSymbolSelectorArgs, AgentRuntimeArgs,
+    AgentSelectorHandle, AgentSymbolArgs, AgentSymbolMode, AgentSymbolViewArgs,
+    AgentWorkspaceFilesArgs, AgentWorkspaceFilesField, AgentWorkspaceFilesViewArgs, KastGraphArgs,
+    KastGraphCommand, KastGraphNodesPageToken, KastPathsArgs, KastRefreshArgs, KastRefreshCommand,
+    KastSymbolArgs, KastSymbolCommand, NativeGraphOperation, OutputFormat, WorkspaceDirtyFilter,
+    WorkspaceFilesPublicPageToken, WorkspaceRelativeGlob,
+};
+use crate::error::{CliError, Result};
+use crate::runtime::{RuntimeState, RuntimeStatusResponse};
+use crate::{config, output, runtime};
+use serde::Serialize;
+use serde_json::{Value, json};
+use std::collections::BTreeSet;
+use std::path::{Path, PathBuf};
+use std::time::{Duration, Instant};
+
+include!("commands.rs");
+include!("refresh.rs");
+include!("graph.rs");
+include!("discovery.rs");
+include!("output.rs");
+include!("external.rs");
+include!("support.rs");
+include!("tests.rs");
