+++ cli-rs/src/main.rs
@@ -1,5 +1,7 @@
 mod agent;
+#[path = "agent/adapter/mod.rs"]
 mod agent_adapter;
+#[path = "agent/plan/mod.rs"]
 mod agent_plan;
 #[path = "configuration/bundle.rs"]
 mod bundle;
@@ -81,6 +83,8 @@ enum Entrypoint {
     Control,
 }
 
+include!("interface/entrypoint/invocation.rs");
+
 fn control_main() -> i32 {
     match parse_cli() {
         Ok(Some(cli)) => {
@@ -121,51 +125,6 @@ fn agent_main() -> i32 {
             error_exit_code(&error)
         }
     }
-}
-
-fn invoked_entrypoint() -> Option<Entrypoint> {
-    match Path::new(&current_executable_argument())
-        .file_stem()
-        .and_then(|name| name.to_str())
-    {
-        Some("kast") => Some(Entrypoint::Agent),
-        Some("_kastctl") => Some(Entrypoint::Control),
-        _ => None,
-    }
-}
-
-fn unrecognized_entrypoint_main() -> i32 {
-    let invoked = Path::new(&current_executable_argument())
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or("unknown")
-        .to_string();
-    let error = CliError::new(
-        "CLI_USAGE",
-        format!(
-            "This executable must be invoked as `kast` or `_kastctl`; `{invoked}` is not supported."
-        ),
-    );
-    let _ = print_agent_error(&error);
-    2
-}
-
-#[derive(Debug, Serialize)]
-struct AgentError<'a> {
-    error: &'a str,
-    message: &'a str,
-    next: &'static str,
-}
-
-fn print_agent_error(error: &CliError) -> Result<()> {
-    output::print_structured(
-        &AgentError {
-            error: error.code,
-            message: &error.message,
-            next: "Run `kast --help` for valid commands and arguments.",
-        },
-        OutputFormat::Toon,
-    )
 }
 
 #[derive(Debug, Serialize)]
