+++ cli-rs/tests/agent/public/kast_resources.rs
